cloudbuild: bookend script with steps to create and destroy cluster
This change makes it so that the script creates and destroys a new
cluster on each run. It uses the `COMMIT_SHA` as the cluster name and
according to https://cloud.google.com/build/docs/configuring-builds/substitute-variable-values#using_default_substitutions
this env variable is present when using the github trigger. For testing
manually, you can manually provide a value for the to-be-created cluster
like this:

```
gcloud builds submit --region=$$REGION \
  --config=cloudbuild-test.yaml \
  --substitutions=COMMIT_SHA=$SHA
```