deployment: make otel container use args
It is common in other OTel Collector k8s deployments to assume that
`CMD` for the container is going to contain the otel collector binary
and to simply apply args. (If you are in CNCF Slack `#otel-operator`
channel, see [this
thread](https://cloud-native.slack.com/archives/C033BJ8BASU/p1728327750985359)).
This PR switches the `opentelemetry-collector` `container` in the
deployment to use args instead of command, and maintain this assumption
about the `CMD` of the container being correct. This means other
properly formed collector container images can be used as a drop-in for
the deployment.